requirements.txt
sqlalchemy
python-telegram-bot
psycopg2
feedparser
